+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The changelog for `Nine41`. Also see the [releases](https://github.com/jessesqui
 
 - Now requires Xcode 11.4
 
+- Add Swift Package Manager support ([@jessesquires](https://github.com/jessesquires)))
+
 - Set date text to `Tuesday January 9`, which shows on iPad status bars ([#4](https://github.com/jessesquires/Nine41/pull/4), [@tfe](https://github.com/tfe))
 
 - Setup SwiftLint for project ([#5](https://github.com/jessesquires/Nine41/issues/5), [@jessesquires](https://github.com/jessesquires))